+++ test/web/activity_pub/mrf/hellthread_policy_test.exs
@@ -26,6 +26,8 @@ defmodule Pleroma.Web.ActivityPub.MRF.HellthreadPolicyTest do
     [user: user, message: message]
   end
 
+  clear_config(:mrf_hellthread)
+
   describe "reject" do
     test "rejects the message if the recipient count is above reject_threshold", %{
       message: message